Split oversized test suites
